% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_with_datestamp.R
\name{knit_with_datestamp}
\alias{knit_with_datestamp}
\title{Knit function to add a datestamp to the name of the Rmd output}
\usage{
knit_with_datestamp(input, ...)
}
\arguments{
\item{input}{A .Rmd file}
}
\description{
Knit function to add a datestamp to the name of the Rmd output
}
\references{
See https://jschultecloos.github.io/reproducr/ for examples.
}
