% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducr_manuscript.R
\name{reproducr_manuscript}
\alias{reproducr_manuscript}
\title{Reproducr manuscript output format based on R Markdown and Bookdown}
\usage{
reproducr_manuscript(
  blinded = FALSE,
  toc = FALSE,
  number_sections = FALSE,
  extra_dependencies = NULL,
  fig_width = 7,
  fig_height = 4,
  fig_crop = TRUE,
  fig_align = "center",
  fig_showtext = TRUE,
  dev = "ragg_png",
  dpi = 500,
  highlight = "default",
  colorlinks = TRUE,
  ...
)
}
\arguments{
\item{blinded}{\code{TRUE} to remove any author information from the title page (defaults to \code{FALSE})}

\item{number_sections}{Whether to number section headers: if \code{TRUE},
figure/table numbers will be of the form \code{X.i}, where \code{X} is the
current first-level section number, and \code{i} is an incremental number
(the i-th figure/table); if \code{FALSE}, figures/tables will be numbered
sequentially in the document from 1, 2, ..., and you cannot cross-reference
section headers in this case.}

\item{fig_showtext}{globally sets the \code{fig.showtext} chunk option (defaults to \code{TRUE})
(defaults to \code{TRUE}). Use the \pkg{showtext} package to rely on custom fonts for
graphs.}

\item{dev}{Graphics device to use for figure output (defaults to 'ragg_png')
from the \pkg{ragg} package}

\item{highlight}{syntax highlighting style in your document (defaults to \code{'tango'}).
Other available code highlighting options are all options currently supported by
\code{rmarkdown::pdf_document}: "pygments", "kate", "monochrome", "espresso",
"zenburn", "haddock", "breezedark". Pass \code{NULL} to prevent syntax highlighting.}

\item{colorlinks}{should cross-references and links be colored (defaults to \code{TRUE})}

\item{...}{Other arguments to be passed to \code{\link{reproducr_manuscript}}}
}
\description{
Creating and publishing reproducible scientific documents in manuscript (PDF) mode
}
\details{
Available metadata variables include:

\describe{
\item{\code{fontsize}}{Font size (e.g. 10pt, 11pt, 12pt)}
\item{\code{documentclass}}{LaTeX document class (currently supported: article or scrartcl)}
\item{\code{classoption}}{Option for \code{documentclass} (e.g. DIV=calc); may be repeated}
\item{\code{mainfont, sansfont, monofont, mathfont}}{Document fonts (works only with xelatex and lualatex, see the \code{latex_engine} option)}
\item{\code{linkcolor, urlcolor, citecolor}}{Color for internal, external, and citation links (red, green, magenta, cyan, blue, black)}
\item{\code{linestretch}}{Options for line spacing (e.g. 1, 1.25, 1.5)}
\item{\code{indent}}{if true, pandoc will use document class settings for indentation (the default LaTeX template otherwise removes indentation and adds space between paragraphs)}
}

\code{reproducr_manuscript()} is a PDF output format function that is optimised for
reproducible scientific research and is based on R Markdown and Bookdown.
Learn more about how to use \code{reproducr_manuscript} here: \link{https://jschultecloos.github.io/reproducr/}.
}
\examples{
\dontrun{
# devtools::install_github("jschultecloos/reproducr")
library(reproducr)
rmarkdown::render("reproducr-file.Rmd", "reproducr::reproducr_manuscript")
}
}
\references{
See https://jschultecloos.github.io/reproducr/ for examples.
}
