% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducr-package.R
\docType{package}
\name{reproducr-package}
\alias{reproducr-package}
\alias{_PACKAGE}
\alias{reproducr}
\title{Drafting and publishing reproducible scientific articles with R Markdown}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The `reproducr` package allows users without any prior knowledge of R Markdown to implement reproducible research practices in their scientific workflows. It provides an integrated Rmd-framework that is fully optimized for the explorative research stage and the research dissemination stage.
}
\details{
The \pkg{reproducr} package allows users without any prior knowledge of R Markdown
to implement reproducible research practices in their scientific workflows.
It provides a single Rmd-template that is fully optimized for two different output formats,
HTML and PDF.
\preformatted{
reproducr_manuscript("researchpaper.Rmd")
reproducr_draft("researchpaper.Rmd")
}
}
\seealso{
The \pkg{reproducr} package builds on the \pkg{bookdown} and \pkg{rmarkdown} package
for the manuscript PDF output.
The \pkg{reproducr} package builds on the \pkg{distill} package
for the draft HTML output.
}
\author{
\strong{Maintainer}: Julia Schulte-Cloos \email{julia.schulte@lmu.de} (\href{https://orcid.org/0000-0001-7223-3602}{ORCID})

}
\keyword{internal}
